refactor: update test-utils to support Lingui
