+++ web-components/src/components/cards/OrderSummaryCard/OrderSummaryCard.test.tsx
@@ -1,6 +1,6 @@
 import { screen } from '@testing-library/dom';
-import { fireEvent, render } from '@testing-library/react';
 import { CURRENCIES } from 'web-components/src/components/DenomIconWithCurrency/DenomIconWithCurrency.constants';
+import { fireEvent, render } from 'web-marketplace/test/test-utils';
 
 import { OrderSummaryCard } from './OrderSummaryCard';
 import { OrderSummaryProps } from './OrderSummaryCard.types';
