+++ web-components/test/test-utils.tsx
@@ -1,0 +1,1 @@
+export * from '../../test/test-utils';
