+++ web-marketplace/vite.config.mts
@@ -31,6 +31,7 @@ export default defineConfig(({ mode }) => {
         config: path.resolve(__dirname, 'src/config'),
         ledger: path.resolve(__dirname, 'src/ledger'),
         clients: path.resolve(__dirname, 'src/clients'),
+        test: path.resolve(__dirname, '../test'),
       },
     },
     build:
